Discrete daemon source
