+++ src/actions.h
@@ -15,17 +15,8 @@ int detach_program(char** argv, enum cli_action act, const other_args* others);
 
 int testrig_ident(other_args* others);
 int testrig_stat(other_args* others);
-int testrig_daemon(other_args* others);
 int testrig_open(other_args* others);
 int testrig_request(other_args* others);
 int testrig_close(other_args* others);
 
 void free_other_args(other_args* others);
-
-enum TESTRIG_DAEMON_STATE {
-	TESTRIG_DAEMON_NONE = 1,
-	TESTRIG_DAEMON_LISTENING = 2,
-	TESTRIG_DAEMON_CLEANING = 4,
-	TESTRIG_DAEMON_STOPPED = 8,
-	TESTRIG_DAEMON_CONNECTED = 16
-};
