+++ src/meson.build
@@ -1,6 +1,6 @@
 cli = executable('testrig',
   dependencies: [ dep ],
-  sources: [ 'main.c', 'args.c', 'actions.c', 'help.c' ],
+  sources: [ 'main.c', 'args.c', 'actions.c', 'help.c', 'daemon.c' ],
   install_dir: where_bin,
   install_rpath: std_rpath,
   install: true)
